requirements.txt
fastapi
python-jose[cryptography]
uvicorn[standard]
requests
pydantic